Fix Duplicate Dependency and Update AWS SD
Fixed a duplicate entry for source-map-support in the dependencies section of package.json. Updated AWS SDK related packages to the latest compatible versions, including @aws-sdk/client-dynamodb, @aws-sdk/client-cognito-identity-provider, and others. Also, ensured the proper versioning for jsonwebtoken, jwk-to-pem, and jwt-decode. This update enhances the security and functionality of the dependencies, and ensures proper structure in the project’s package configuration.